this is 2nd commit

+++ Java_Challage/src/main/java/com/project/controller/AccountController.java
@@ -1,14 +1,10 @@
 package com.project.controller;
 
-import java.io.FileReader;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 
-import org.json.simple.JSONArray;
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,8 +15,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.project.dto.AccountDTO;
+import com.project.enums.ExceptionEnum;
+import com.project.exception.CustomException;
 import com.project.model.Account;
-import com.project.model.PersonDetail;
 import com.project.service.AccountService;
 import com.project.service.PersonDetailService;
 
@@ -33,37 +30,42 @@ public class AccountController {
 	@Autowired
 	private PersonDetailService personDetailService;
 
+	//TODO: save acoount
 	@PostMapping(value = "/saveAccount")
-	public ResponseEntity saveAccount(@RequestBody AccountDTO accountDto) throws Exception {
-		HttpStatus status;
+	public ResponseEntity saveAccount(@Valid@RequestBody AccountDTO accountDto) throws Exception {
 		try {
+			
 			this.accountService.saveAccount(accountDto);
-			status=HttpStatus.OK;
+
 		} catch (Exception e) {
-			status=HttpStatus.INTERNAL_SERVER_ERROR;
 			throw e;
 		}
-		return new ResponseEntity<>(status);
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
 
+	//TODO: view all account
 	@GetMapping(value = "/viewAccounts")
 	public ResponseEntity getAllAccount() {
 		List<Account> sccountList = new ArrayList<>();
-		HttpStatus status;
 		
 		try {
 			sccountList = this.accountService.getAllAccount();
-			status=HttpStatus.OK;
 		} catch (Exception e) {
-			status=HttpStatus.INTERNAL_SERVER_ERROR;
 			throw e;
 		}
-		return new ResponseEntity<>(sccountList, status);
+		return new ResponseEntity<>(sccountList, HttpStatus.OK);
 	}
-
-	@PostMapping(value = "/saveAndUpdate")
+	
+	//TODO: update account
+	@PostMapping(value = "/updateAndSave")
 	public ResponseEntity saveAndUpdate(@RequestBody AccountDTO accDto, @RequestParam Integer accountId)
 			throws Exception {
+		
+		if(accountId  ==  null) {
+			throw new CustomException(ExceptionEnum.INVALID_ID.getMessage(), 
+					ExceptionEnum.INVALID_ID.getStatusCode());
+		}
+		
 		try {
 			this.accountService.updateAccount(accDto, accountId);
 
@@ -72,11 +74,22 @@ public class AccountController {
 			throw e;
 		}
 	}
-
-	@GetMapping(value = "/deleteAccounts")
-	public ResponseEntity deleteAccounts(@RequestParam int accountId) {
-
-		this.accountService.deleteAccount(accountId);
+	
+	//TODO: delete account
+	@GetMapping(value = "/deleteAccount")
+	public ResponseEntity deleteAccount(@RequestParam Integer accountId) {
+		
+		try {
+			if(accountId  ==  null) {
+				throw new CustomException(ExceptionEnum.INVALID_ID.getMessage(), 
+						ExceptionEnum.INVALID_ID.getStatusCode());
+			}
+			
+			this.accountService.deleteAccount(accountId);
+		}catch (Exception e) {
+			// TODO: handle exception
+			throw e;
+		}
 
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
